+++ example/deployment/heroku/app.py
@@ -7,7 +7,7 @@ import os
 import websockets
 
 
-async def echo(websocket, path):
+async def echo(websocket):
     async for message in websocket:
         await websocket.send(message)