+++ example/shutdown_server.py
@@ -4,7 +4,7 @@ import asyncio
 import signal
 import websockets
 
-async def echo(websocket, path):
+async def echo(websocket):
     async for message in websocket:
         await websocket.send(message)